oak_loader: add placeholder links to FFI crate
Add an `ffi_glue` feature to the `oak_loader` crate, which expects to
link to the `oak_eulg` crate.

From Cargo, leave this feature turned off (i.e. not in default_features)
so a normal `cargo build` does not require FFI linkage.

From Bazel, turn this feature on and so trigger the placeholder links
across to C++.

While we're adding features, also add an `oak_debug` feature to the
`oak_loader` crate which governs whether logging, metrics and
introspection are enabled.