runtime: use Weak refs for GC and introspection
Move `ChannelMapping` to just hold a `Weak<Channel>` rather than a full
`Arc`, so dropping `ChannelHalf` objects doesn't need special processing
to allow for the extra strong reference.

However, this does mean that each `Channel` needs to be able to get back
to the `ChannelMapping`, so when it is dropped it can remove itself from
the mapping.